History

+++ src/Controller/HomeSectionController.php
@@ -36,7 +36,7 @@ class HomeSectionController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="new", methods={"GET","POST"})
+     * @Route("/section/new", name="section_new", methods={"GET","POST"})
      * @param Request $request
      * @param SluggerInterface $slugger
      * @return Response
@@ -80,7 +80,7 @@ class HomeSectionController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="show", methods={"GET"})
+     * @Route("/section/show/{id}", name="section_show", methods={"GET"})
      * @param Section $section
      * @return Response
      */
@@ -92,7 +92,7 @@ class HomeSectionController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/edit", name="edit", methods={"GET","POST"})
+     * @Route("/section/edit/{id}", name="section_edit", methods={"GET","POST"})
      * @param Request $request
      * @param Section $section
      * @return Response
@@ -115,7 +115,7 @@ class HomeSectionController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="delete", methods={"DELETE"})
+     * @Route("/section/{id}", name="section_delete", methods={"DELETE"})
      * @param Request $request
      * @param Section $section
      * @param EntityManagerInterface $entityManager
@@ -125,12 +125,27 @@ class HomeSectionController extends AbstractController
     {
         if ($this->isCsrfTokenValid('delete' . $section->getId(), $request->request->get('_token'))) {
             $filename = $section->getImage();
-            $path = $this->getParameter('upload_dir') . '/' . $filename;
-            unlink($path);
+            if (is_string($this->getParameter('upload_dir'))) {
+                $path = $this->getParameter('upload_dir') . $filename;
+                unlink($path);
+            }
+
             $entityManager->remove($section);
             $entityManager->flush();
         }
 
         return $this->redirectToRoute('admin_homeSection');
+    }
+
+    /**
+     * @Route("/section/sport/{id}", name="section_sport", methods={"GET"})
+     * @param Section $section
+     * @return Response
+     */
+    public function sport(Section $section): Response
+    {
+        return $this->render('section/sports/sport.html.twig', [
+            'section' => $section,
+        ]);
     }
 }
